% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_child_species.R
\name{get_child_species}
\alias{get_child_species}
\title{Returns a list of species that are children of the target species}
\usage{
get_child_species(target, species)
}
\arguments{
\item{target}{A \code{git_repository} object to store the imported data.}
}
\description{
Returns a list of species that are children of the target species
}
