% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_id.R
\name{get_project_id_hibernation}
\alias{get_project_id_hibernation}
\title{Get project id for the monitoring of hibernating bats}
\usage{
get_project_id_hibernation(origin)
}
\arguments{
\item{origin}{A DBI connection to the database}
}
\description{
Get project id for the monitoring of hibernating bats
}
