% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_relevant_visit.R
\name{list_relevant_visit}
\alias{list_relevant_visit}
\title{List relevant visits for hibernation analysis}
\usage{
list_relevant_visit(data_root, n_winter = 24)
}
\arguments{
\item{n_winter}{The number of winters to consider for relevance.}
}
\value{
A data frame with the relevant visits and their relevance status.
}
\description{
We apply the following rules to determine if a visit is relevant location by
location:
A relevant visit is no older than \code{n_winter} winters.
When a location is divided into sublocations, we count the number of visits
with observations at the sublocation level.
If there are at least three such winters with visits where bats are observed,
and the number of winters with detailed observations is larger than the
number with only total counts, only the visits with detailed observation are
relevant.
Otherwise, we look at all visits, and if there are at least three winters
with visits where bats are observed, then all visits within the last
\code{n_winter} are relevant.
}
