% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_model_species.R
\name{prepare_analysis_model_species}
\alias{prepare_analysis_model_species}
\title{Prepare the analysis models for a single species}
\usage{
prepare_analysis_model_species(
  raw_data,
  species,
  base,
  project = "batanalysis",
  n_extrapolation = 24,
  max_dist = 10,
  visits,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{raw_data}{The git repository with the raw data}

\item{species}{the code of the species}

\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{n_extrapolation}{Impute only locations or sublocations when there is no
more than \code{n_extrapolation} winters between the observation to impute and the
nearest observation.}

\item{max_dist}{The maximum distance for the range in kilometres}

\item{visits}{The output of \code{visit_type()}}

\item{overwrite}{Should an existing object be overwritten?
Defaults to \code{TRUE}.}

\item{verbose}{A logical.
When \code{TRUE} print the message.
When \code{FALSE} do nothing.}
}
\description{
Prepare the analysis models for a single species
}
