% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_imputation_species.R
\name{prepare_imputation_species}
\alias{prepare_imputation_species}
\title{Prepare the imputation model for a species}
\usage{
prepare_imputation_species(
  target,
  species = 545,
  start = as.Date("2000-07-01")
)
}
\arguments{
\item{target}{A \code{git_repository} object to store the imported data.}

\item{species}{the ID of the species}

\item{start}{The oldest date to use in the analysis}
}
\description{
Prepare the imputation model for a species
}
