% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cluster.R
\name{read_cluster}
\alias{read_cluster}
\title{Import object clusters}
\usage{
read_cluster(cluster, target)
}
\arguments{
\item{cluster}{the name of the comma separated cluster file}

\item{target}{A \code{git_repository} object to store the imported data.}
}
\description{
Import object clusters
}
