% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_individual.R
\name{read_raw_individual}
\alias{read_raw_individual}
\title{Import the raw observations according to the individual based protocol}
\usage{
read_raw_individual(origin, ignore = c("dead", "flying"))
}
\arguments{
\item{origin}{A \code{DBI} connection to the SQL Server database.}

\item{ignore}{A character vector with the activities to ignore.
Defaults to \code{c("dead", "flying")}.}
}
\description{
Import the raw observations according to the individual based protocol
}
