% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_location.R
\name{read_raw_location}
\alias{read_raw_location}
\title{Import the location list}
\usage{
read_raw_location(origin)
}
\arguments{
\item{origin}{A \code{DBI} connection to the SQL Server database.}
}
\description{
Import the location list
}
