% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_species.R
\name{read_raw_species}
\alias{read_raw_species}
\title{Import the species list}
\usage{
read_raw_species(origin)
}
\arguments{
\item{origin}{A \code{DBI} connection to the SQL Server database.}
}
\description{
Import the species list
}
