% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_total.R
\name{read_raw_total}
\alias{read_raw_total}
\title{Import the raw observations according to the totals based protocol}
\usage{
read_raw_total(origin, ignore = c("dead", "flying"))
}
\arguments{
\item{origin}{A \code{DBI} connection to the SQL Server database.}

\item{ignore}{A character vector with the activities to ignore.
Defaults to \code{c("dead", "flying")}.}
}
\value{
A list with the visits, observations, and problems.
}
\description{
Import the raw observations according to the totals based protocol
}
