% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_imputation_detail.R
\name{select_imputation_detail}
\alias{select_imputation_detail}
\title{Select detailed data for imputation}
\usage{
select_imputation_detail(
  locations,
  raw_data,
  this_species,
  start,
  n_winter = 4,
  n_present = 3,
  max_delta = 10,
  n_extrapolation = 5
)
}
\arguments{
\item{locations}{A data frame with location information.
Must contain the columns \code{location_id}, \code{detailed} and \code{type}.}

\item{raw_data}{The git repository with the raw data}

\item{this_species}{A data frame with one ore more relevant species ID.}

\item{start}{The oldest date to use in the analysis.}

\item{n_winter}{Minimum number of winters in which a (sub)location is
monitored.}

\item{n_present}{Minimum number of winters in which the species is observed
at a (sub)location.}

\item{max_delta}{Only relevant in case of multiple visits within the same
winter to a location divided into sublocation.
We select for every winter per sublocation the visit that is closest to the
middle of the winter (15th of January).
All selected visits of a location for a given winter should be within
\code{max_delta} days from each other.
If not, we drop the visits the furthest away from the middle of the winter
until the condition is met.
This ensures that the visits to the sublocations are not too far apart in
time.
The default is 10 days.
This means that all visits to sublocations of a location should be
within 10 days from each other.
This allows to spread the visits over a few days.}

\item{n_extrapolation}{Maximum number of winters from the nearest winter
where the species was observed at a (sub)location.
Only do imputations for a (sub)location when the difference between the
nearest observed winter and the missing winter is below or equal to this
threshold.}
}
\description{
Select detailed data for imputation
}
