% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_imputation_no_detail.R
\name{select_imputation_no_detail}
\alias{select_imputation_no_detail}
\title{Select imputation data from non-detailed locations}
\usage{
select_imputation_no_detail(
  locations,
  raw_data,
  this_species,
  start,
  n_winter = 4,
  n_present = 3,
  n_extrapolation = 5
)
}
\arguments{
\item{locations}{A data frame with location information.
Must contain the columns \code{location_id}, \code{detailed} and \code{type}.}

\item{raw_data}{The git repository with the raw data}

\item{this_species}{A data frame with one ore more relevant species ID.}

\item{start}{The oldest date to use in the analysis.}

\item{n_winter}{Minimum number of winters in which a (sub)location is
monitored.}

\item{n_present}{Minimum number of winters in which the species is observed
at a (sub)location.}

\item{n_extrapolation}{Maximum number of winters from the nearest winter
where the species was observed at a (sub)location.
Only do imputations for a (sub)location when the difference between the
nearest observed winter and the missing winter is below or equal to this
threshold.}
}
\description{
Select imputation data from non-detailed locations
}
