% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_imputation_species.R
\name{select_imputation_species}
\alias{select_imputation_species}
\title{Select the relevant data for the imputation model for a species}
\usage{
select_imputation_species(
  target,
  species = 545,
  start = as.Date("2000-07-01"),
  n_winter = 2
)
}
\arguments{
\item{target}{A \code{git_repository} object to store the imported data.}

\item{species}{the ID of the species}

\item{start}{The oldest date to use in the analysis}

\item{n_winter}{Minimum number of winters in which the species is observed at
a location.
Remove locations with a number below this threshold.}
}
\description{
Select the relevant data for the imputation model for a species
}
