% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_badge.R
\name{survey_badge}
\alias{survey_badge}
\title{Survey badge}
\usage{
survey_badge(target, duration = 24)
}
\arguments{
\item{target}{A \code{git_repository} object to store the imported data.}

\item{duration}{The number of recent winters to consider for the badge.}
}
\value{
A data frame with the location ID, badge type, and score.
}
\description{
This function calculates a survey badge based on the number of visits to bat
monitoring locations.
The badge looks at the usable visits over the last 24 winters.
A visit is usable if it has either a total count or section counts.
If a location has sublocations, only visits with section counts are counted.
The score is calculated as a weighted sum of the visits, with more recent
visits having a higher weight.
The weights are calculated as \code{1 / n}, where \code{n} is the number of winters
ago the visit took place.
We rescale the weights so that the maximum score is 1.
The badge is then assigned based on the score:
\itemize{
\item Gold: a score equal to or higher than the score of having a visit in
the 75\% most recent winters.
\item Silver: a score equal to or higher than the score of having a visit in
the 50\% most recent winters.
\item Bronze: a score equal to or higher than the score of having a visit in
the 25\% most recent winters.
\item None: a score lower than the score of having a visit in the 25\% most
recent winters.
}
}
