% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsurveyed_locations.R
\name{unsurveyed_sublocations}
\alias{unsurveyed_sublocations}
\title{Count the unsurveyed sublocations per location}
\usage{
unsurveyed_sublocations(
  data_repo = keyring::key_get("meetnetten", "batanalysis_data")
)
}
\arguments{
\item{data_repo}{Character; path to the root of the data repository}
}
\value{
A tibble with columns location_id, location_name, sublocation_id,
sublocation, n (number of visits where the sublocation was not surveyed)
}
\description{
Count the unsurveyed sublocations per location
}
