% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_tile.R
\name{osm_tile}
\alias{osm_tile}
\title{return OSM raster tiles for a given bounding box}
\usage{
osm_tile(
  bbox = c(xmin = 4.26, ymin = 51.22, xmax = 4.5, ymax = 51.35),
  extra_zoom = 0,
  cache_path = R_user_dir("batanalysis", which = "cache"),
  type = c("osm", "water")
)
}
\arguments{
\item{bbox}{Either (i) four numeric values specifying the maximal and minimal
longitudes and latitudes, in the form \code{c(xmin, ymin, xmax, ymax)}
or (ii) a character string in the form \code{xmin,ymin,xmax,ymax}. These
will be passed to \link[osmdata]{getbb} to be converted to a numerical bounding
box. Can also be (iii) a matrix representing a bounding polygon as
returned from \code{getbb(..., format_out = "polygon")}. To search in an
area, (iv) a character string with a relation or a (closed) way id in
the format \code{"way(id:1)"}, \code{"relation(id:1, 2)"} or \code{"relation(id:1, 2, 3); way(id:2)"} as returned by \code{getbb(..., format_out = "osm_type_id")}
or \link[osmdata]{bbox_to_string} with a \code{data.frame} from \code{getbb(..., format_out = "data.frame")} to select all areas combined (relations and ways).}

\item{extra_zoom}{the number of extra zoom levels}

\item{cache_path}{File location to cache maps}

\item{type}{Character; type of map to download.
Options are \code{"osm"} (OpenStreetMap, default) and
\code{"water"} (watercolor style from Stadia Maps)}
}
\description{
return OSM raster tiles for a given bounding box
}
