% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_tile.R
\name{osm_tile}
\alias{osm_tile}
\title{return OSM raster tiles for a given bounding box}
\usage{
osm_tile(
  bbox = c(xmin = 4.26, ymin = 51.22, xmax = 4.5, ymax = 51.35),
  extra_zoom = 0,
  cache_path = R_user_dir("batanalysis", which = "cache"),
  type = c("osm", "water")
)
}
\arguments{
\item{extra_zoom}{the number of extra zoom levels}

\item{cache_path}{File location to cache maps}
}
\description{
return OSM raster tiles for a given bounding box
}
