% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_data.R
\name{prepare_analysis_data}
\alias{prepare_analysis_data}
\title{Convert raw data to analysis data for all species}
\usage{
prepare_analysis_data(
  raw_data,
  analysis_data,
  start,
  n_winter = 4,
  n_present = 3
)
}
\arguments{
\item{raw_data}{A gity repository with the raw data.}

\item{analysis_data}{A gity repository to store the analysis data.}

\item{start}{The oldest date to use in the analysis}

\item{n_winter}{Minimum number of winters in which the species is observed at
a location.
Remove locations with a number below this threshold.}
}
\description{
Convert raw data to analysis data for all species
}
