% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_model.R,
%   R/prepare_model_analysis.R
\name{prepare_analysis_model}
\alias{prepare_analysis_model}
\title{Prepare analysis models for all species}
\usage{
prepare_analysis_model(
  analysis_data,
  base,
  max_dist = 10,
  project = "batanalysis",
  overwrite = FALSE
)

prepare_analysis_model(
  analysis_data,
  base,
  max_dist = 10,
  project = "batanalysis",
  overwrite = FALSE
)
}
\arguments{
\item{analysis_data}{A gity repository to store the analysis data.}

\item{base}{The base location to store the model.}

\item{max_dist}{The maximum distance for the range in kilometers}

\item{project}{Will be a relative path within the base location.}

\item{overwrite}{Should an existing object be overwritten?
Defaults to \code{TRUE}.}
}
\description{
Prepare analysis models for all species

Prepare analysis models for all species
}
