% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_model.R
\name{prepare_analysis_model}
\alias{prepare_analysis_model}
\title{Prepare analysis models for all species}
\usage{
prepare_analysis_model(
  raw_data,
  start_winter = as.integer(format(Sys.Date(), "\%Y")) - 23,
  base,
  project = "batanalysis",
  max_delta = 10,
  n_extrapolation = 24,
  max_dist = 10,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{raw_data}{The git repository with the raw data}

\item{base}{The base location to store the model.}

\item{project}{Will be a relative path within the base location.}

\item{max_dist}{The maximum distance for the range in kilometres}

\item{overwrite}{Should an existing object be overwritten?
Defaults to \code{TRUE}.}

\item{verbose}{A logical.
When \code{TRUE} print the message.
When \code{FALSE} do nothing.}
}
\description{
Prepare analysis models for all species
}
