% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_individual.R
\name{read_raw_individual}
\alias{read_raw_individual}
\title{Import the raw observations according to the individual based protocol}
\usage{
read_raw_individual(origin)
}
\arguments{
\item{origin}{A \code{DBI} connection to the SQL Server database.}
}
\description{
Import the raw observations according to the individual based protocol
}
