% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_section.R
\name{read_raw_section}
\alias{read_raw_section}
\title{Import the raw observations according to the section based protocol}
\usage{
read_raw_section(origin)
}
\arguments{
\item{origin}{A \code{DBI} connection to the SQL Server database.}
}
\description{
Import the raw observations according to the section based protocol
}
