% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_imputation_total.R
\name{select_imputation_total}
\alias{select_imputation_total}
\title{Select the relevant data for the imputation model for a species by total.}
\usage{
select_imputation_total(
  target,
  species = "Mmysbra",
  start = Sys.Date() - 12 * 365
)
}
\arguments{
\item{target}{A \code{git_repository} object to store the imported data.}

\item{species}{the ID of the species}

\item{start}{The oldest date to use in the analysis}
}
\description{
Use only the data collected using the totals based protocol.
}
